refactor(orchestration): extract HumanDelegationService
The last leaf service. requestDelegateForHuman (207L) and validateRequest (21L)
move verbatim, along with the unexported free function isRequestDelegateInput.
827 -> 589 lines. Twelve units now exist.

The plan contradicted itself here. Its Method Assignment gives requestDelegate to
this service; its Task 19 note says the three overload signatures must stay on the
facade so router-types.ts resolves OrchestrationService["requestDelegate"]; and
Task 20 requires every facade method body to be at most three lines, which the
four-line if/else dispatcher is not. Splitting signature from body satisfies all
three: the facade keeps the overloads with a one-line delegation, the service
carries the same overloads and the real dispatcher. tsc is the check -- a wrong
facade signature fails in router-types.ts, not here.

isRequestDelegateInput moves into the service file rather than being imported back
from the facade. It is a value, so verbatimModuleSyntax would not erase that
import, and the facade imports the service class as a value -- a real runtime
cycle. Tasks 12, 13 and 18 each had to route around the same trap.

The plan's port list named config, ensureWorkerSession and logger; none of the
three methods uses any of them. Five ports.

claimParallelStart is called inside a critical section, releaseParallelStart and
the reserve/ensure/dispatch calls outside. Unchanged.
